+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 ### Changed
 - When a user is blocked, any DM streams between the blocker and blocked user are torn down.
 - Add confirm dialog to console delete operations.
+- Reduce Console Storage View memory usage.
 
 ### Fixed
 - Ensure Apple receipts with duplicate transaction identifiers are processed cleanly.